Message compression implementation is no longer super janky.
My previous approach to message compression was: only parse the message once, character-by-character. But message compression requires re-examining previously-parsed characters! I decided to achieve these contradictory goals by building up a "context" object, which was just a map of offsets to domain names. Then, when the parser encountered a pointer, it could simply look up the pointer's value in the map.

But this requires managing a lot of state. Building the map is hard, because it requires every parser to know which offset it's currently at (so that it can accurately report the offset of any labels it discovers).

Instead, my new approach is that the Parser object always has two copies of the input. One is a slice being parsed and consumed byte-by-byte. The other is immutable, always contains the full message, and is used to produce new slices at whichever offset a pointer points at.

This new approach is much simpler to implement and has trivial runtime penalty. I suspect there's a way to use copy-on-write to avoid having a second copy of the DNS message response in memory, but hey, they're only 512 bytes, who really cares?